pyproject.toml
[project]
name = "nguyenpanda-config"
version = "0.1.0"
description = "A dynamic YAML configuration engine supporting variable interpolation, function calls, and file inheritance."
authors = [
    { name = "Tuong Nguyen Ha", email = "hatuongnguyen@0107gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "nguyenpanda>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/nguyenpanda/PandaConfig"
# Documentation = ""
# Changelog = ""
Source = "https://github.com/nguyenpanda/PandaConfig"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "deepdiff>=8.6.1",
    "pytest>=9.0.2",
    "twine>=6.2.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
]
